% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down.ifdm.R
\name{down.ifdm}
\alias{down.ifdm}
\title{down.ifdm}
\usage{
down.ifdm(indicador = "Geral")
}
\arguments{
\item{indicador}{pode receber um dos quatro valores
Geral - indice geral do idfm
Educ - indicador de educacao
Empre - indicador de Emprego e Renda
Saud - indicador de Saude}
}
\description{
down.ifdm
Faz o download dos indicadores do Indice Firjan de Desenvolvimento Municipal (ifdm)
O resuldo eh um arquivo em xls no diretorio de trabalho
a serie disponivel eh de 2003 a 2013, eh um arquivo único
}
\examples{
\dontrun{
down.ifdm("Geral")
}
}
