% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ifdm.R
\name{read.ifdm}
\alias{read.ifdm}
\title{read.ifdm Faz a leitura dos arquivos (em xls)indicadores do Indice Firjan de Desenvolvimento Municipal (ifdm)
O resuldo eh um data.frame}
\usage{
read.ifdm(arquivo, linha = F, nota = NULL)
}
\arguments{
\item{arquivo}{pode receber o endereco do arquivo}

\item{linha}{se FALSE (padrao) retorna um data.frame com os indicadores e ranking por ano em colunas,
se TRUE retorna um data.frame com os indicadores ou ranking por ano em linhas (observacao)}

\item{nota}{se TRUE retorna a nota do indicador se, FALSE ternona o ranking,
para esse paremetro o paremtro linha tem ser TRUE.}
}
\description{
read.ifdm
Faz o retorna um data.frame dos indicadores do Indice Firjan de Desenvolvimento Municipal (ifdm)
O resuldo eh um arquivo em xls no diretorio de trabalho
a serie disponivel eh de 2003 a 2013, eh um arquivo único
}
\examples{
\dontrun{
read.ifdm("Geral_ifdm.xls")
}

}
